2009-03-20  Philip Van Hoof  <pvanhoof@codeminded.be>
        * daemon/plugins/gstreamer-video-plugin.c: Seeking 30 seconds into the
        video before making a thumbnail of the video.

        Fixes Bug# 92748

        * daemon/hildon-thumbnail-plugin.h
        * daemon/plugins/gdkpixbuf-png-out-plugin.c
        * daemon/plugins/gdkpixbuf-jpeg-out-plugin.c
        * daemon/thumbnailer.c
        * daemon/thumbnailer.h
        * daemon/thumbnailer.xml
        * daemon/hildon-thumbnail-plugin.c: DBus API change for the 'since'
        parameter of cleanup



git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk/hildon-thumbnail@17763 50eb543b-83fa-0310-9bb7-f7cbbee4ab39

+++ daemon/hildon-thumbnail-plugin.h
@@ -76,7 +76,7 @@ gboolean    hildon_thumbnail_outplugins_needs_out (HildonThumbnailPluginOutType 
 gchar *     hildon_thumbnail_outplugins_get_orig  (const gchar *path);
 
 gchar *     hildon_thumbnail_outplugins_cleanup   (const gchar *uri_match, 
-						   guint64 since);
+						   guint since);
 
 
 G_END_DECLS
